Rename channel helper reducers
