+++ packages/wallet/src/redux/reducers/__tests__/running.test.ts
@@ -1,4 +1,4 @@
-import { runningReducer } from '../channels/running';
+import { runningReducer } from '../channels/runningReducer';
 import * as scenarios from './test-scenarios';
 import * as states from '../../channel/state';
 import * as actions from '../../actions';
