+++ packages/wallet/src/redux/reducers/channels/closingReducer.ts
