+++ packages/wallet/src/redux/reducers/channels/openingReducer.ts
